Layout 컴포넌트 생성하여 theme provider 적용
